Upgraded SQL handler

+++ src/main/java/com/synload/framework/http/UploadedFile.java
@@ -2,7 +2,7 @@ package com.synload.framework.http;
 
 public class UploadedFile {
 	public String name, path, tempName, user = "";
-	public long size = (long)0;
+	public long size = 0;
 	public UploadedFile(String name, String path, String tempName, String user, long size){
 		this.size = size;
 		this.name = name;
